Convert to AgentCheck decorator
