+++ datadog_checks_base/datadog_checks/base/checks/base.py
@@ -2,6 +2,7 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import copy
+import functools
 import importlib
 import inspect
 import json
@@ -606,6 +607,31 @@ class AgentCheck(object):
     def is_metadata_collection_enabled():
         # type: () -> bool
         return is_affirmative(datadog_agent.get_config('enable_metadata_collection'))
+
+    @classmethod
+    def metadata_entrypoint(cls, method):
+        # type: (Callable[..., None]) -> Callable[..., None]
+        """
+        Mark a method as a metadata entrypoint.
+
+        This decorator provides:
+
+        * Automatic no-op behavior in case metadata collection is disabled on the Agent.
+        * Automatic exception handling.
+        """
+
+        @functools.wraps(method)
+        def entrypoint(self, *args, **kwargs):
+            # type: (AgentCheck, *Any, **Any) -> None
+            if not self.is_metadata_collection_enabled():
+                return
+
+            try:
+                method(self, *args, **kwargs)
+            except Exception as exc:
+                self.log.debug('Error collecting metadata: %s', exc)
+
+        return entrypoint
 
     def set_external_tags(self, external_tags):
         # type: (List[ExternalTagType]) -> None
