+++ datadog_checks_base/datadog_checks/base/utils/metadata/__init__.py
@@ -2,4 +2,3 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .core import MetadataManager
-from .decorators import metadata_entrypoint
